feat(badges): add badges to (admin) user profiles
