+++ app/controllers/admin/users_controller.rb
@@ -19,6 +19,8 @@ class Admin::UsersController < AdminController
 
     @team_memberships = @user.team_memberships(false)
     @link_to_admin_events = true
+    
+    @header_badges = helpers.user_labels_for(@user, nil)
 
     super
   end
