+++ app/helpers/user_helper.rb
@@ -1,0 +1,49 @@
+module UserHelper
+    # Generate the labels to be shown in the team member list & user profile.
+    # Deadline is the proposal deadline, if applicable.
+    def team_member_labels_for(team_member, deadline)
+        output_labels = []
+
+        show_member_status =  (team_member.teamwork_type == "Event" && team_member.teamwork.this_academic_year?) || (deadline.present? && deadline.future?)
+
+        output_labels << { label_class: "bg-secondary", text: "Not A Member" } if !team_member.user.has_role?("Member") && show_member_status
+
+        output_labels + user_labels_for(team_member.user, deadline)
+    end
+
+    # Generate the labels to be shown in the team member list & user profile.
+    # Deadline is the proposal deadline, if applicable.
+    def user_labels_for(user, deadline)
+        output_labels = []
+
+        user.roles.trained.each do |role|
+            label_class = role.name == "First Aid Trained" ? "bg-success" : "bg-info"
+            output_labels << { label_class: label_class, text: role.name }
+        end
+
+        if user.has_role?("life member")
+            output_labels << { label_class: "bg-rainbow-rotate", text: "Life Member" }
+        end
+
+
+        # Display the 'Not a Member' label if the show is this academic year, or it has a deadline in the future (and is likely a proposal)
+
+        if user.in_debt
+            debt_message = user.debt_message_suffix.upcase_first
+
+            output_labels << { label_class: "bg-danger", text: link_to(debt_message, admin_debt_path(user)) }
+        end
+
+        if deadline.present? && user.in_debt(deadline)
+            on_deadline_debt_message = user.debt_message_suffix(deadline).upcase_first
+
+            # Only show the deadline message if the debt on the deadline is different.
+            if on_deadline_debt_message != debt_message
+                on_deadline_debt_message = "#{on_deadline_debt_message} on the editing deadline"
+                output_labels << { label_class: "bg-danger", text: link_to(on_deadline_debt_message, admin_debt_path(user)) }
+            end
+        end
+
+        output_labels
+    end
+end
