+++ app/views/shared/_admin_page_header.html.erb
@@ -1,10 +1,15 @@
-<%# locals: (title:) -%>
+<%# locals: (title:, header_badges: nil) -%>
+
+<% header_badges = [] if header_badges.nil? %>
 
 <div class="content-header">
     <div class="container-fluid">
         <div class="row mb-2">
-            <div class="col">
+            <div class="flex-grow-1 ml-2">
                 <h1 class="m-0"><%= title %></h1>
+                <% header_badges.each do |label| %>
+                  <%= generate_label(label[:label_class], label[:text]) %>
+                <% end %>
             </div><!-- /.col -->
             <div class="col-auto">
                 <ol class="breadcrumb float-sm-right">
